+++ packages/expo-updates/ios/EXUpdates/Update/EXUpdatesNewUpdate.m
@@ -11,32 +11,34 @@ NS_ASSUME_NONNULL_BEGIN
 
 @implementation EXUpdatesNewUpdate
 
-+ (EXUpdatesUpdate *)updateWithNewManifest:(NSDictionary *)rootManifest
++ (EXUpdatesUpdate *)updateWithNewManifest:(EXUpdatesNewRawManifest *)rootManifest
                                   response:(nullable NSURLResponse *)response
                                     config:(EXUpdatesConfig *)config
                                   database:(EXUpdatesDatabase *)database
 {
-  NSDictionary *manifest = rootManifest;
-  if (manifest[@"manifest"]) {
-    manifest = manifest[@"manifest"];
+  NSDictionary *tempManifest = rootManifest.rawManifestJSON;
+  if (tempManifest[@"manifest"]) {
+    tempManifest = tempManifest[@"manifest"];
   }
-
+  
+  EXUpdatesNewRawManifest *manifest = [[EXUpdatesNewRawManifest alloc] initWithRawManifestJSON:tempManifest];
+  
   EXUpdatesUpdate *update = [[EXUpdatesUpdate alloc] initWithRawManifest:manifest
                                                                   config:config
                                                                 database:database];
-
-  id updateId = manifest[@"id"];
-  id commitTime = manifest[@"createdAt"];
-  id runtimeVersion = manifest[@"runtimeVersion"];
-  id launchAsset = manifest[@"launchAsset"];
-  id assets = manifest[@"assets"];
-
+  
+  id updateId = manifest.rawID;
+  id commitTime = manifest.createdAt;
+  id runtimeVersion = manifest.runtimeVersion;
+  id launchAsset = manifest.launchAsset;
+  id assets = manifest.assets;
+  
   NSAssert([updateId isKindOfClass:[NSString class]], @"update ID should be a string");
   NSAssert([commitTime isKindOfClass:[NSString class]], @"createdAt should be a string");
   NSAssert([runtimeVersion isKindOfClass:[NSString class]], @"runtimeVersion should be a string");
   NSAssert([launchAsset isKindOfClass:[NSDictionary class]], @"launchAsset should be a dictionary");
   NSAssert(!assets || [assets isKindOfClass:[NSArray class]], @"assets should be null or an array");
-
+  
   NSUUID *uuid = [[NSUUID alloc] initWithUUIDString:(NSString *)updateId];
   NSAssert(uuid, @"update ID should be a valid UUID");
   
@@ -44,16 +46,16 @@ NS_ASSUME_NONNULL_BEGIN
   NSAssert([bundleUrlString isKindOfClass:[NSString class]], @"launchAsset.url should be a string");
   NSURL *bundleUrl = [NSURL URLWithString:bundleUrlString];
   NSAssert(bundleUrl, @"launchAsset.url should be a valid URL");
-
+  
   NSMutableArray<EXUpdatesAsset *> *processedAssets = [NSMutableArray new];
-
+  
   NSString *bundleKey = launchAsset[@"key"];
   EXUpdatesAsset *jsBundleAsset = [[EXUpdatesAsset alloc] initWithKey:bundleKey type:EXUpdatesEmbeddedBundleFileType];
   jsBundleAsset.url = bundleUrl;
   jsBundleAsset.isLaunchAsset = YES;
   jsBundleAsset.mainBundleFilename = EXUpdatesEmbeddedBundleFilename;
   [processedAssets addObject:jsBundleAsset];
-
+  
   if (assets) {
     for (NSDictionary *assetDict in (NSArray *)assets) {
       NSAssert([assetDict isKindOfClass:[NSDictionary class]], @"assets must be objects");
@@ -67,24 +69,24 @@ NS_ASSUME_NONNULL_BEGIN
       NSAssert(type && [type isKindOfClass:[NSString class]], @"asset contentType should be a nonnull string");
       NSURL *url = [NSURL URLWithString:(NSString *)urlString];
       NSAssert(url, @"asset url should be a valid URL");
-
+      
       EXUpdatesAsset *asset = [[EXUpdatesAsset alloc] initWithKey:key type:(NSString *)type];
       asset.url = url;
-
+      
       if (metadata) {
         NSAssert([metadata isKindOfClass:[NSDictionary class]], @"asset metadata should be an object");
         asset.metadata = (NSDictionary *)metadata;
       }
-
+      
       if (mainBundleFilename) {
         NSAssert([mainBundleFilename isKindOfClass:[NSString class]], @"asset localPath should be a string");
         asset.mainBundleFilename = (NSString *)mainBundleFilename;
       }
-
+      
       [processedAssets addObject:asset];
     }
   }
-
+  
   update.updateId = uuid;
   update.commitTime = [RCTConvert NSDate:(NSString *)commitTime];
   update.runtimeVersion = (NSString *)runtimeVersion;
@@ -92,14 +94,14 @@ NS_ASSUME_NONNULL_BEGIN
   update.keep = YES;
   update.bundleUrl = bundleUrl;
   update.assets = processedAssets;
-  update.metadata = manifest;
-
+  update.metadata = manifest.rawManifestJSON;
+  
   if ([response isKindOfClass:[NSHTTPURLResponse class]]) {
     NSDictionary *headersDictionary = ((NSHTTPURLResponse *)response).allHeaderFields;
     update.serverDefinedHeaders = [[self class] dictionaryWithStructuredHeader:headersDictionary[@"expo-server-defined-headers"]];
     update.manifestFilters = [[self class] dictionaryWithStructuredHeader:headersDictionary[@"expo-manifest-filters"]];
   }
-
+  
   return update;
 }
 
@@ -108,7 +110,7 @@ NS_ASSUME_NONNULL_BEGIN
   if (!headerString) {
     return nil;
   }
-
+  
   EXStructuredHeadersParser *parser = [[EXStructuredHeadersParser alloc] initWithRawInput:headerString fieldType:EXStructuredHeadersParserFieldTypeDictionary ignoringParameters:YES];
   NSError *error;
   NSDictionary *parserOutput = [parser parseStructuredFieldsWithError:&error];
@@ -116,7 +118,7 @@ NS_ASSUME_NONNULL_BEGIN
     NSLog(@"Error parsing header value: %@", error ? error.localizedDescription : @"Header was not a structured fields dictionary");
     return nil;
   }
-
+  
   NSMutableDictionary *mutableDict = [NSMutableDictionary dictionaryWithCapacity:parserOutput.count];
   [parserOutput enumerateKeysAndObjectsUsingBlock:^(id key, id obj, BOOL *stop) {
     // ignore any dictionary entries whose type is not string, number, or boolean